requirements.txt
# Core dependencies
streamlit>=1.28.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# PyTorch for segmentation model
torch>=2.0.0
torchvision>=0.15.0

# TensorFlow for classification model
tensorflow>=2.13.0
keras>=2.13.0

# YOLO detection
ultralytics>=8.0.0

# RAG system dependencies
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU available
python-dotenv>=1.0.0

# FEMA 306 RAG dependencies
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5

# Utilities
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Interactive canvas for brick calibration
streamlit-image-coordinates>=0.1.6
